+++ src/ui/units/ql/utils/visualization/getAvailableQlVisualizations.ts
@@ -1,8 +1,7 @@
 import cloneDeep from 'lodash/cloneDeep';
-import merge from 'lodash/merge';
+import {ChartkitGlobalSettings, Placeholder, Shared} from 'shared';
+import {DL} from 'ui/constants/common';
 
-import {ChartkitGlobalSettings, Placeholder, Shared} from '../../../../shared';
-import {DL} from '../../../constants';
 import {
     AREA_100P_VISUALIZATION,
     AREA_VISUALIZATION,
@@ -20,8 +19,7 @@ import {
     SCATTER_D3_VISUALIZATION,
     SCATTER_VISUALIZATION,
     TREEMAP_VISUALIZATION,
-} from '../../../constants/visualizations';
-import {DEFAULT_VISUALIZATION_ID_QL} from '../constants';
+} from '../../../../constants/visualizations';
 
 export function getAvailableQlVisualizations(
     options?: ChartkitGlobalSettings,
@@ -115,29 +113,3 @@ export function getAvailableQlVisualizations(
         },
     );
 }
-
-export function getDefaultQlVisualization(): Shared['visualization'] {
-    const availableVisualizations = getAvailableQlVisualizations();
-
-    // We use column chart as initial visualization type in QL and Wizard
-    return (
-        availableVisualizations.find(
-            (visualization) => visualization.id === DEFAULT_VISUALIZATION_ID_QL,
-        ) || availableVisualizations[0]
-    );
-}
-
-export const getQlVisualization = (
-    visualizationId: string,
-    loadedVisualization: Shared['visualization'],
-): Shared['visualization'] => {
-    const availableVisualizations = getAvailableQlVisualizations();
-
-    const candidate = availableVisualizations.find((vis) => vis.id === visualizationId);
-
-    if (candidate) {
-        return merge({}, candidate, loadedVisualization);
-    } else {
-        return getDefaultQlVisualization();
-    }
-};
